added ParityRPC class, moved parity related functions from eth.ts
